refactor: remove commands.ts re-exports, import directly from submodules
